Update uiowa_auth README (#5274)
This module is not currently installed via Composer so this line should be removed. Future installation instructions should be provided when that process is clarified. It might be possible to use Composer in combination with TravisCI release uploading for uiowa/uiowa releases. For reference: https://github.com/uiowa/uiowa/issues/2630